Added directories and changed namespaces; Added MessageQueueOptions; Refactored queue to use MessageQueueOptions
