+++ Source/Microsoft.Teams.Apps.CompanyCommunicator.Common/Services/MessageQueues/DataQueue/DataQueue.cs
@@ -2,9 +2,9 @@
 // Copyright (c) Microsoft. All rights reserved.
 // </copyright>
 
-namespace Microsoft.Teams.Apps.CompanyCommunicator.Common.Services.MessageQueue
+namespace Microsoft.Teams.Apps.CompanyCommunicator.Common.Services.MessageQueues.DataQueue
 {
-    using Microsoft.Extensions.Configuration;
+    using Microsoft.Extensions.Options;
 
     /// <summary>
     /// The message queue service connected to the "company-communicator-data" queue in Azure service bus.
@@ -19,9 +19,11 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Common.Services.MessageQueue
         /// <summary>
         /// Initializes a new instance of the <see cref="DataQueue"/> class.
         /// </summary>
-        /// <param name="configuration">ASP.NET Core <see cref="IConfiguration"/> instance.</param>
-        public DataQueue(IConfiguration configuration)
-            : base(configuration, DataQueue.QueueName)
+        /// <param name="messageQueueOptions">The message queue options.</param>
+        public DataQueue(IOptions<MessageQueueOptions> messageQueueOptions)
+            : base(
+                  serviceBusConnectionString: messageQueueOptions.Value.ServiceBusConnection,
+                  queueName: DataQueue.QueueName)
         {
         }
     }
