+++ Source/Microsoft.Teams.Apps.CompanyCommunicator.Common/Services/MessageQueues/DataQueue/DataQueueMessageContent.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // </copyright>
 
-namespace Microsoft.Teams.Apps.CompanyCommunicator.Common.Services.MessageQueue
+namespace Microsoft.Teams.Apps.CompanyCommunicator.Common.Services.MessageQueues.DataQueue
 {
     using System;
