+++ Source/Microsoft.Teams.Apps.CompanyCommunicator.Common/Services/MessageQueues/PrepareToSendQueue/PrepareToSendQueueMessageContent.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // </copyright>
 
-namespace Microsoft.Teams.Apps.CompanyCommunicator.Common.Services.MessageQueue
+namespace Microsoft.Teams.Apps.CompanyCommunicator.Common.Services.MessageQueues.PrepareToSendQueue
 {
     /// <summary>
     /// Azure service bus prepare to send queue message content class.
